+++ andes/routines/tds.py
@@ -39,7 +39,8 @@ class TDS(BaseRoutine):
                                      ('qrt', 0),
                                      ('kqrt', 1.0),
                                      ('store_f', 0.0),
-                                     ('store_g', 0.0),
+                                     ('store_h', 0.0),
+                                     ('store_i', 0.0),
                                      )))
         self.config.add_extra("_help",
                               tol="convergence tolerance",
@@ -56,7 +57,8 @@ class TDS(BaseRoutine):
                               qrt='quasi-real-time stepping',
                               kqrt='quasi-real-time scaling factor; kqrt > 1 means slowing down',
                               store_f='store RHS of diff. equations',
-                              store_g='store RHS of algebraic equations',
+                              store_h='store RHS of external diff. equations',
+                              store_i='store RHS of external algeb. equations',
                               )
         self.config.add_extra("_alt",
                               tol="float",
@@ -73,7 +75,8 @@ class TDS(BaseRoutine):
                               qrt='bool',
                               kqrt='positive',
                               store_f=(0, 1),
-                              store_g=(0, 1),
+                              store_h=(0, 1),
+                              store_i=(0, 1),
                               )
 
         # overwrite `tf` from command line
@@ -156,6 +159,7 @@ class TDS(BaseRoutine):
         system.dae.clear_ts()
         system.store_sparse_pattern(models=system.exist.pflow_tds)
         system.store_adder_setter(models=system.exist.pflow_tds)
+        system.store_no_check_init(models=system.exist.pflow_tds)
         system.vars_to_models()
 
         system.init(system.exist.tds, routine='tds')
@@ -302,14 +306,16 @@ class TDS(BaseRoutine):
 
             if step_status:
                 f_vals = dae.f if self.config.store_f else None
-                g_vals = dae.g if self.config.store_g else None
+                h_vals = dae.h if self.config.store_h else None
+                i_vals = dae.i if self.config.store_i else None
 
                 dae.ts.store(dae.t.tolist(),
                              x=dae.x,
                              y=dae.y,
                              z=system.get_z(models=system.exist.pflow_tds),
                              f=f_vals,
-                             g=g_vals,
+                             h=h_vals,
+                             i=i_vals,
                              )
 
                 self.streaming_step()
@@ -675,6 +681,9 @@ class TDS(BaseRoutine):
         system = self.system
         self.fg_update(system.exist.pflow_tds)
         system.j_update(models=system.exist.pflow_tds)
+
+        # reset diff. RHS where `check_init == False`
+        system.dae.f[system.no_check_init] = 0.0
 
         # warn if variables are initialized at limits
         if system.config.warn_limits: