+++ docs/source/release-notes.rst
@@ -14,6 +14,9 @@ v1.3.0
 - Allow `State` variable set `check_init=False` to skip
   initialization test. One use case is for integrators
   with non-zero inputs (such as state-of-charge integration).
+- Added `Jumper` for connecting two buses with zero impedance.
+- `REGCA1` and synchronous generators can take power ratio
+  parameters `gammap` and `gammaq`.
 
 v1.2 Notes
 ----------