+++ docs/source/release-notes.rst
@@ -17,7 +17,7 @@ v1.3.0
 - Added `Jumper` for connecting two buses with zero impedance.
 - `REGCA1` and synchronous generators can take power ratio
   parameters `gammap` and `gammaq`.
-- New models: `IEESGO`.
+- New models: `IEESGO` and `IEEET1`.
 
 v1.2 Notes
 ----------